# src/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Car:
    properties:
      available:
        type: boolean
      basicInfo:
        type: string
      brand:
        type: string
      carPics:
        items:
          $ref: '#/definitions/models.CarsPic'
        type: array
      carType:
        type: string
      comments2car:
        items:
          $ref: '#/definitions/models.Comment2Car'
        type: array
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      featureInfo:
        type: string
      id:
        type: integer
      location:
        type: string
      model:
        type: string
      numSeats:
        type: integer
      ownerId:
        type: integer
      price:
        type: number
      rating:
        type: number
      supportDelivery:
        type: boolean
      supportDriver:
        type: boolean
      updatedAt:
        type: string
      useLogs:
        items:
          $ref: '#/definitions/models.UserCar'
        type: array
      year:
        type: integer
    type: object
  models.CarsPic:
    properties:
      carId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      fileName:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.Comment2Car:
    properties:
      bookId:
        type: integer
      carId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      likes:
        type: integer
      mainContent:
        type: string
      photos:
        items:
          $ref: '#/definitions/models.Comment2CarPic'
        type: array
      titleContent:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  models.Comment2CarPic:
    properties:
      comment2CarId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      fileName:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.Comment2User:
    properties:
      bookId:
        type: integer
      carId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      likes:
        type: integer
      mainContent:
        type: string
      photos:
        items:
          $ref: '#/definitions/models.Comment2UserPic'
        type: array
      titileContent:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  models.Comment2UserPic:
    properties:
      comment2CarId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      fileName:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.FavoriteCars:
    properties:
      carId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
...
  models.User:
    properties:
      comments2user:
        items:
          $ref: '#/definitions/models.Comment2User'
        type: array
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      favoriteCars:
        items:
          $ref: '#/definitions/models.FavoriteCars'
        type: array
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      myCars:
        items:
          $ref: '#/definitions/models.Car'
        type: array
      password:
        type: string
      updatedAt:
        type: string
      userPic:
        type: string
      username:
        type: string
    type: object
  models.UserCar:
    properties:
      UserId:
        type: integer
      carId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      endTime:
        type: string
      id:
        type: integer
      reason:
        description: |-
          0: Request     1:Pending for payment     2:Paid but not used
          3.During using 4.After using             5.Finished
        type: string
      startTime:
        type: string
      status:
        type: integer
      totalPrice:
        type: integer
      updatedAt:
        type: string
    type: object
  models.UserProfile:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      myCars:
        items:
          $ref: '#/definitions/models.Car'
        type: array
      updatedAt:
        type: string
      userPic:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: cheayuki13@gmail.com
  description: This is a car rental system API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Car Rental System API
  version: "1.0"
paths:
...
  /api/v1/user/getProfile/{id}:
    get:
      consumes:
      - application/json
      description: Get the profile of a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserProfile'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user profile
      tags:
      - User
...
  /api/v1/user/register:
    post:
      consumes:
      - application/json
      description: New user to register
      parameters:
      - description: User Registration Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: User Register
      tags:
      - User
  /api/v1/user/updateUserProfile/{id}:
    post:
      consumes:
      - application/json
      description: Update the profile of a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User object that needs to be updated
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user profile
      tags:
      - User
swagger: "2.0"